Add working optionals for runtime and shutdown
The runtime function pointer is called by the FreeSWITCH core if
present causing an extra useless thread to be launched. Now the
optional function pointers back for the runtime and shutdown functions
but the static table defaults to None and during module load we
check if the module definition set valid functions and we install
our wrapper if we have a valid function definition.

This was not strictly needed for the shutdown pointer but keeps
things cleaner.